requirements.txt
pandas
mysql-connector-python
streamlit
openpyxl
tqdm
seaborn
matplotlib
